automation_framework/requirements.txt
allure-pytest
allure-python-commons
playwright
pytest
pytest-playwright
PyYAML
requests
pytest-html
pytest-metadata
pytest-xdist
PyMySQL
cryptography
tenacity
black
ruff